% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BlackBoxNCV.R
\name{BlackBoxNCV}
\alias{BlackBoxNCV}
\title{Nested cross fold validation with BlackBox}
\usage{
BlackBoxNCV(data, labels, outerfolds, innerfolds, ntrees, mTry, Kernel, Gamma,
  exclude, inn.exclude, Method, AUC, metric)
}
\arguments{
\item{data}{A data.frame where the columns correspond to features and the rows are samples. The dataframe will be shuffled and split into k folds for downstream analysis.}

\item{labels}{A character or numeric vector of the class indetifiers that each sample belongs.}

\item{outerfolds}{The number of folds that will be in the first k-fold loop, this determines the number of holdouts.}

\item{innerfolds}{The number of folds that occur in the internal feature selection cross fold validation before testing on the corresponding holdout.}

\item{ntrees}{The number of trees used in the ensemble based learners (randomforest, bigrf, party, bartmachine). default = 500.}

\item{mTry}{The number of features sampled at each node in the trees of ensemble based learners (randomforest, bigrf, party, bartmachine). default = sqrt(number of features).}

\item{Kernel}{The type of kernel used in the support vector machine algorithm (linear, radial, sigmoid, polynomial). default = "linear".}

\item{Gamma}{Advanced parameter, defines the distance of which a single training example reaches. Low gamma will produce a SVM with softer boundries, as Gamma increases the boundries will eventually become restricted to their singular support vector.}

\item{exclude}{removes certain algorithms from analysis - to exclude random forest which you would set exclude = c(1). To only run GLM you would set exclude = c(1:4,6:8). The algorithms each have their own numeric identifier. randomforest = 1, knn = 2, bartmachine = 3, party = 4, glm = 5, pam = 6, nnet = 7, svm = 8.}

\item{inn.exclude}{removes certain algorithms from after feature selection analysis. similar to 'exclude'.}

\item{Method}{The algorithm used to feature select the data. Uses the feature importance from the algorithms to rank and remove anything below the AUC threshold.}

\item{AUC}{Area under the curve selection measure. The relative importance of features is calculated and then ranked. The features responsible for the most importance are therefore desired, the AUC value is the percentile in which to keep features above. 0.5 keeps the highest ranked features responsible for 50 percent of the cumulative importance.}

\item{metric}{A character string to determine which performance metric will be passed on to the Performance() function. Refer to Perfromance() documentation.}

\item{seeds}{A numeric vector}
}
\description{
A function that builds upon the BlackBox and BlackBoxNCV function and performs nested k-fold cross validation and then provides votes for each fold as well as the importance of each feature in the models.
}
\author{
Zachary Davies, Boris Guennewig
}
\keyword{AUC,}
\keyword{BlackBox,}
\keyword{Cross}
\keyword{Validation,}
\keyword{feature}
\keyword{k-fold,}
\keyword{selection}

