% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/blkboxROC.R
\name{blkboxROC}
\alias{blkboxROC}
\title{ROC plots for blkbox}
\usage{
blkboxROC(results, title)
}
\arguments{
\item{results}{The output of blkbox Performance that had "AUROC" as one of the specified metrics.}

\item{title}{The title of the plot. Default is "ROC".}
}
\description{
will plot ROC curves for output from Performance function if "AUROC" was specified.
}
\author{
Zachary Davies, Boris Guennewig
}
\keyword{AUC,}
\keyword{Cross}
\keyword{Validation,}
\keyword{blkbox,}
\keyword{feature}
\keyword{k-fold,}
\keyword{selection,}

