% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/blkbox.R
\name{blkbox}
\alias{blkbox}
\title{Train and Test datasets.}
\usage{
blkbox(data, labels, holdout, holdout.labels, ntrees, mTry, Kernel, Gamma,
  exclude)
}
\arguments{
\item{data}{a data frame of training data where the features correspond to columns and the samples are rows. As data size increases the memory required and run time of some algorithms may compound exponentially.}

\item{holdout}{a data frame of holdout of testing data where the features correspond to columns and the samples are the rows.}

\item{ntrees}{The number of trees used in the ensemble based learners (randomforest, bigrf, party, bartmachine). default = 500.}

\item{mTry}{The number of features sampled at each node in the trees of ensemble based learners (randomforest, bigrf, party, bartmachine). default = sqrt(number of features).}

\item{Kernel}{The type of kernel used in the support vector machine algorithm (linear, radial, sigmoid, polynomial). default = "linear".}

\item{exclude}{removes certain algorithms from analysis - to exclude random forest which you would set exclude = c(1). To only run GLM you would set exclude = c(1:4,6:8). The algorithms each have their own numeric identifier. randomforest = 1, knn = 2, bartmachine = 3, party = 4, glm = 5, pam = 6, nnet = 7, svm = 8.}

\item{a}{character or numeric vector that contains the training class identifiers for the samples in the data frame. Must appear in the same order.}

\item{a}{character or numeric vector that contains the holdout or testing class identifiers for the samples in the holdout data frame.}

\item{Advanced}{parameter, defines the distance of which a single training example reaches. Low gamma will produce a SVM with softer boundaries, as Gamma increases the boundaries will eventually become restricted to their singular support vector. default is 1/(ncol - 1).}
}
\description{
This standard function will allow multiple machine learning algorithms to be utilized on the same data to determine, which algorithm may be the most appropriate.
}
\examples{
\donttest{
library(blkbox)
options(java.parameters = "-Xmx8000m" )
bb_res = blkbox(data = , labels = , holdout = , holdout.labels = , exclude = c(1:7))
Performance(bb_res, metric = "AUROC", consensus = T)
}
}
\author{
Zachary Davies, Boris Guennewig
}
\keyword{Learning,}
\keyword{Machine}
\keyword{Testing}
\keyword{Training,}
\keyword{blkbox,}

