% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blkboxROC.R
\name{blkboxROC}
\alias{blkboxROC}
\title{ROC plots for blkbox}
\usage{
blkboxROC(results, title = "ROC")
}
\arguments{
\item{results}{The output of blkbox Performance that had "AUROC" as one of the specified metrics.}

\item{title}{The title of the plot. Default is "ROC".}
}
\description{
will plot ROC curves for output from Performance function if "AUROC" was specified.
}
\examples{
\donttest{
# model_1 can be any blkbox or blkboxCV model
perf = Performance(model_1)
# Standard ROC curve
blkboxROC(perf)
}
}
\author{
Zachary Davies, Boris Guennewig
}
\keyword{ROC,}
\keyword{blkbox}

