% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/blkboxVenn.R
\name{cv.venn}
\alias{cv.venn}
\title{Crossfold Validation Venn Diagrams.}
\usage{
cv.venn(results)
}
\arguments{
\item{results}{An object produced by the blkboxCV function.}
}
\description{
This relies on numerous methods being used for feature selection in blkboxCV. It will construct a venn diagram to compare the features selected by various algorithms based on feature importance.
}
\author{
Zachary Davies, Boris Guennewig
}
\keyword{CV}
\keyword{Venn,}
\keyword{Venndiagram,}
\keyword{blkbox,}
\keyword{crossfold,}

