% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BlackBoxVenn.R
\name{ncv.venn}
\alias{ncv.venn}
\title{Nested Crossfold Validation Venn Diagrams}
\usage{
ncv.venn(results)
}
\arguments{
\item{results}{An object produced by the blkboxNCV function.}
}
\description{
Compares the features used in the holdouts of nested cross fold validation, this will tell you how much variataion in features selected there is between holdouts.
}
\author{
Zachary Davies, Boris Guennewig
}
\keyword{NCV}
\keyword{Nested,}
\keyword{Venn,}
\keyword{Venndiagram,}
\keyword{blkbox,}

